Add starter code
Add the StockHawk app code to the repo.  Use a branch to associate each
push with an issue identified by number and description.  The starter
issues have been added already.  Other will be added as required.

Also added a Gradle snippet to allow the source code (but not the build
artifacts) to be stored on Google Drive to facilitate easy code sharing
among multiple computers.